+++ src/instruction.rs
@@ -256,6 +256,10 @@ impl Info for Kind {
             _ => false,
         }
     }
+
+    fn ignored() -> Self {
+        Self::nop
+    }
 }
 
 impl fmt::Display for Kind {
@@ -325,7 +329,7 @@ impl From<Size> for u64 {
 }
 
 /// A single RISC-V instruction
-#[derive(Copy, Clone, Default, Debug, Eq, PartialEq)]
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub struct Instruction<I: Info = Option<Kind>> {
     /// [`Size`] of the instruction
     pub size: Size,
@@ -351,6 +355,12 @@ impl<I: Info> Instruction<I> {
         let size = bits.size();
         let info = base.decode_bits(bits);
         Self { size, info }
+    }
+}
+
+impl<I: Info> Default for Instruction<I> {
+    fn default() -> Self {
+        Info::ignored()
     }
 }
 
@@ -411,6 +421,13 @@ impl<I: Info> Info for Instruction<I> {
 
     fn is_return(&self) -> bool {
         self.info.is_return()
+    }
+
+    fn ignored() -> Self {
+        Self {
+            info: I::ignored(),
+            size: Default::default(),
+        }
     }
 }